refactor(hickey): move resolvedPr to dockModel.ts, delete dockRowChrome
After the palette unification, `dockRowChrome.ts` had a single
remaining export (`resolvedPr`) — a pure narrower on
`TerminalMetadata["pr"]` with zero dock-row dependency. Three
consumers (`Dock.tsx`, `MobileDockDrawer.tsx`, `commands.ts`)
imported it; for `commands.ts` that meant the app-level palette
registry reaching into a `canvas/dock/` chrome file for a metadata
helper — a layering inversion.

`dockModel.ts` already wraps `TerminalMetadata` for workspace
search (`prSearchFields` packs the same union into the search
corpus), so co-locating the resolver removes the misnamed module
without spawning a new single-function file.
